backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1
httpx==0.28.1
anthropic==0.42.0
PyJWT[crypto]==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
